Add links to specs for each property
This fetches https://www.w3.org/Style/CSS/all-properties.en.json, (via a
proxy for CORS reasons,) which provides a link to each spec that defines
a given property. It creates a list of links to the specs for each
property. It's a really ugly list, because I don't have a sass compiler,
but it's a list nonetheless.

I assumed the JSON data would be identical to the list here:
https://www.w3.org/Style/CSS/all-properties.en.html ...but I was
mistaken. It includes whatwg specs too, which we probably want to filter
out to reduce the noise.

Anyway, I'm tired. Maybe this is useful to you! Maybe not, IDK. XD