Download CSS properties JSON and deploy to GitHub pages
This removes the use of https://api.allorigins.win/ which regularly goes
down and takes (on my fast network) over 30 seconds to load the a 22kb
JSON file! Instead, the site is now deployed to GitHub pages once a
month. As part of the deploy steps the all-properties.en.json is fetched
and saved to the assets folder (which avoids any CORS issues, and is
nice and fast).